car_dealer_app/appointment_service/requirements.txt
Flask
gunicorn
google-cloud-datastore
flask-cors